Update EventEmitter tests to reflect recent changes
